// src/auth/roles/types/role-subjects.type.ts
export enum RoleSubjects {
  User = 'User',
  Quote = 'Quote',
  Vendo = 'Vendo',
  RecentlyViewedQuotesVendos = 'RecentlyViewedQuotesVendos',
  WorkflowRule = 'WorkflowRule',
  VacationRule = 'VacationRule',
  QuoteApprovalQueue = 'QuoteApprovalQueue',
  VendoApprovalQueue = 'VendoApprovalQueue',
  CompanyChannel = 'CompanyChannel',
  Company = 'Company',
  Role = 'Role',
  All = 'all', // reserved for casl's all subjects
}